extract class to dedicated file
